+++ lib/ferrum/browser.rb
@@ -29,7 +29,7 @@ module Ferrum
     delegate %i[default_context] => :contexts
     delegate %i[targets create_target page pages windows workers service_workers attach_target] => :default_context
     delegate %i[go_to goto go back forward refresh reload stop wait_for_reload
-                at_css at_xpath css xpath current_url current_title url title
+                at_css at_xpath css xpath wait_for_selector current_url current_title url title
                 body doctype content=
                 headers cookies network accessibility downloads
                 mouse keyboard