+++ lib/ferrum/page.rb
@@ -31,7 +31,7 @@ module Ferrum
 
     extend Forwardable
 
-    delegate %i[at_css at_xpath css xpath
+    delegate %i[at_css at_xpath css xpath wait_for_selector
                 current_url current_title url title body doctype content=
                 execution_id execution_id! evaluate evaluate_on evaluate_async execute evaluate_func
                 add_script_tag add_style_tag] => :main_frame